requirements.txt
attrs==21.2.0
click==8.0.1
DateTime==4.3
Flask==2.0.1
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
packaging==21.0
pluggy==0.13.1
psycopg2-binary==2.9.1
py==1.10.0
pymongo==3.12.0
pyparsing==2.4.7
pytest==6.2.4
python-dotenv==0.19.0
pytz==2021.1
toml==0.10.2
Werkzeug==2.0.1
zope.interface==5.4.0
